// apps/server/functions/src/app/graphql.schema.interface.ts
export interface Address {
    __typename?: 'Address';
    streetAddress: string;
    city: string;
    state: string;
    zipcode: string;
}

export interface Comment {
    __typename?: 'Comment';
    userId: string;
    comment: string;
    created: string;
}

export interface Activity {
    __typename?: 'Activity';
    userId: string;
    activity: string;
    created: string;
}